add ws_futures, refactor and add tests
