+++ binance/ws/reconnecting_websocket.py
@@ -23,7 +23,7 @@ class ReconnectingWebsocket:
     MAX_RECONNECTS = 5
     MAX_RECONNECT_SECONDS = 60
     MIN_RECONNECT_WAIT = 0.1
-    TIMEOUT = 5
+    TIMEOUT = 10
     NO_MESSAGE_RECONNECT_TIMEOUT = 60
     MAX_QUEUE_SIZE = 100
